refactor(core): split agent client internals
